+++ components/tabs/__tests__/Tabs.test.tsx
@@ -48,7 +48,7 @@ describe('Tabs', () => {
     expect(wrapperSmall.classes()).toContain('hmfw-tabs-small')
 
     const wrapperMiddle = mount(Tabs, { props: { items, size: 'middle' } })
-    expect(wrapperMiddle.classes()).not.toContain('hmfw-tabs-middle')
+    expect(wrapperMiddle.classes()).toContain('hmfw-tabs-middle')
 
     const wrapperLarge = mount(Tabs, { props: { items, size: 'large' } })
     expect(wrapperLarge.classes()).toContain('hmfw-tabs-large')